requirements.txt
streamlit
tensorflow-cpu==2.10.0
numpy
Pillow
gdown
streamlit>=1.28.0
tensorflow>=2.10.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
pygame>=2.0.0